refactor: 모델 파일명 변경
