+++ frontend/src/components/home/AvailableDogCheck.tsx
@@ -1,7 +1,7 @@
 import Avatar from '@/components/common/Avatar';
 import { Checkbox } from '@/components/common/Checkbox';
 import { Divider } from '@/components/common/Divider';
-import { AvailableDog } from '@/models/dog.model';
+import { AvailableDog } from '@/models/dog';
 
 export const AvailableDogCheck = ({ dog, onToggle }: { dog: AvailableDog; onToggle: (id: number) => void }) => {
     return (
