+++ frontend/src/components/home/AvailableDogCheckList.tsx
@@ -2,7 +2,7 @@ import AllDogs from '@/assets/icons/walk/frame-5058.svg';
 import Avatar from '@/components/common/Avatar';
 import { Checkbox } from '@/components/common/Checkbox';
 import { AvailableDogCheck } from '@/components/home/AvailableDogCheck';
-import { AvailableDog } from '@/models/dog.model';
+import { AvailableDog } from '@/models/dog';
 import { useState } from 'react';
 
 export default function AvailableDogCheckList({
