+++ frontend/src/components/journals/CustomCalendar.tsx
@@ -5,7 +5,7 @@ import { useLocation } from 'react-router-dom';
 import { queryStringKeys } from '@/constants';
 import PrevMonth from '@/assets/icons/btn-prev-month.svg';
 import NextMonth from '@/assets/icons/btn-next-month.svg';
-import { fetchDogMonthStatistic, period } from '@/api/dogs';
+import { fetchDogMonthStatistic, period } from '@/api/dog';
 import { formDate, formDay } from '@/utils/date';
 import useCalendar from '@/hooks/useCalendar';
