+++ frontend/src/components/walk/DogFeceAndUrineCheckList.tsx
@@ -3,7 +3,7 @@ import { Checkbox } from '@/components/common/Checkbox';
 import { Divider } from '@/components/common/Divider';
 import { Feces } from '@/components/icon/Feces';
 import { Urine } from '@/components/icon/Urine';
-import { WalkingDog } from '@/models/dog.model';
+import { WalkingDog } from '@/models/dog';
 
 interface DogFeceAndUrineCheckListProps {
     dog: WalkingDog;
