+++ frontend/src/components/walk/Map.tsx
@@ -1,7 +1,7 @@
 import Spinner from '@/components/common/Spinner';
 import { DEFAULT_LAT, DEFAULT_LNG } from '@/constants/location';
 import { NAV_HEIGHT, TOP_BAR_HEIGHT, WALK_INFO_HEIGHT } from '@/constants/style';
-import { Position } from '@/models/location.model';
+import { Position } from '@/models/location';
 import { cn } from '@/utils/tailwindClass';
 import { useEffect, useState } from 'react';
