+++ frontend/src/pages/Journals/index.tsx
@@ -4,10 +4,10 @@ import JournalCardList from '@/components/journals/JournalCardList';
 import { NAV_HEIGHT, TOP_BAR_HEIGHT } from '@/constants/style';
 import useJournals from '@/hooks/useJournals';
 import Ic from '@/assets/icons/ic-arrow.svg';
-import { Dog } from '@/models/dog.model';
+import { Dog } from '@/models/dog';
 import { useLocation, useNavigate, useSearchParams } from 'react-router-dom';
 import { useEffect, useState } from 'react';
-import { fetchDogList } from '@/api/dogs';
+import { fetchDogList } from '@/api/dog';
 import Avatar from '@/components/common/Avatar';
 import SelectBox from '@/components/common/SelectBox';
 import { queryStringKeys } from '@/constants';
