+++ frontend/src/pages/Walk.tsx
@@ -15,8 +15,8 @@ import { useLocation, useNavigate } from 'react-router-dom';
 import useToast from '@/hooks/useToast';
 import useStopAlert from '@/hooks/useStopAlert';
 import { getStorage, removeStorage, setStorage } from '@/utils/storage';
-import { WalkingDog } from '@/models/dog.model';
-import { Position } from '@/models/location.model';
+import { WalkingDog } from '@/models/dog';
+import { Position } from '@/models/location';
 import { storageKeys } from '@/constants';
 import { walkStartRequest, walkStopRequest } from '@/api/walk';
 import useImageUpload from '@/hooks/useImageUpload';
